WebOSTVService: Resubscribe subscriptions after foregrounding
Save current subscriptions when the app is entering background, and
automagically resubscribe to them when entering foreground. It is
important that the `ServiceSubscription` objects are the same, because
the app is likely to retain them to be able to unsubscribe.

Unfortunately, functional tests for this change are missing, because
they would be pretty long and incomprehensible (the architecture is not
modularized enough).